requirements.txt
flask==2.0.1
flask-cors==3.0.10
torch==1.9.0
torchvision==0.10.0
opencv-python==4.5.3.56
numpy==1.21.2
pillow==8.3.1
matplotlib==3.4.3
werkzeug==2.0.1
gunicorn==20.1.0
openai~=1.0
python-dotenv~=1.0.0
